api_yamdb/requirements.txt
asgiref==3.2.10
requests~=2.26
django~=2.2
djangorestframework~=3.12.4
PyJWT~=2.1
django-filter~=2.4.0
chardet~=5.0
djangorestframework-simplejwt~=4.8.0
gunicorn==20.0.4
psycopg2-binary==2.9.0
pytz==2020.1
sqlparse==0.3.1
python-dotenv~=0.21
pytest==6.2.4
pytest-django==4.5.2
pytest-pythonpath==0.7.3
